Moved props within component
